get rid of root user finder, finish basic auth implementation, only use validators for extractors they make sense for
