+++ src/plone.server/plone/server/auth/users.py
@@ -1,8 +1,33 @@
 
 
-class RootUserIdentifier(object):
-    def __init__(self, request):
-        self.request = request
+ROOT_USER_ID = 'root'
 
-    async def get_user(self, token):
-        return self.request.application.root_user
+
+class RootUser(object):
+    def __init__(self, password):
+        self.id = ROOT_USER_ID
+        self.password = password
+        self.groups = ['Managers']
+        self._roles = {}
+        self._properties = {}
+
+
+class PloneUser(object):
+
+    def __init__(self, request):
+        self.id = 'plone'
+        self.request = request
+        self._groups = []
+        self._roles = {}
+        self._properties = {}
+
+    @property
+    def groups(self):
+        return self._groups
+
+
+class AnonymousUser(PloneUser):
+
+    def __init__(self, request):
+        super(AnonymousUser, self).__init__(request)
+        self.id = 'Anonymous User'
