+++ src/plone.server/plone/server/factory.py
@@ -4,8 +4,8 @@ from pkg_resources import iter_entry_points
 from plone.server import app_settings
 from plone.server.async import IAsyncUtility
 from plone.server.auth.validators import hash_password
-from plone.server.auth.participation import ROOT_USER_ID
-from plone.server.auth.participation import RootUser
+from plone.server.auth.users import ROOT_USER_ID
+from plone.server.auth.users import RootUser
 from plone.server.content import IStaticDirectory
 from plone.server.content import IStaticFile
 from plone.server.content import StaticFile
